+++ HISTORY.rst
@@ -5,7 +5,11 @@ Release History
 
 0.8.0
 +++++++++++++++
+<<<<<<< HEAD
 * Added support for iot central commands, monitor-events and device-twin show
+=======
+* Added Azure IoT Plug & Play public preview functionality
+>>>>>>> master
 
 0.7.1
 +++++++++++++++